+++ py-polars/src/series.rs
@@ -500,7 +500,7 @@ impl PySeries {
         self.series.arg_max()
     }
 
-    pub fn take(&self, indices: Wrap<AlignedVec<u32>>) -> PyResult<Self> {
+    pub fn take(&self, indices: Wrap<Vec<u32>>) -> PyResult<Self> {
         let indices = indices.0;
         let indices = UInt32Chunked::new_from_aligned_vec("", indices);